requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
ta>=0.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
line-bot-sdk>=2.0.0
python-binance>=1.0.15 